requirements.txt
langchain>=0.1.0
langchain-community
langchain-google-genai
streamlit>=1.25.0
gradio>=3.40.0
python-dotenv
requests
beautifulsoup4
duckduckgo-search
pytest
pyyaml